added support for multi output loss + train
made changes to following main files:

* collate_fn -> support for ((depth, action), keypoint) and (depth, (keypoint, action)) outputs just before passing to model i.e.
		final stage data loader.
* data_loader -> support for multi in single out and single in multi out as described above using config see conf
* data_transformers -> similar support for the ToTuple extractor
* loss+metrics -> suport to extract relevant output and targets as required for the relevant loss function this is a crude way of
		  doing things atm.
* deep-prior-baseline-model -> support for multi-out during forward, multi-in supported but just commented here see new branch.
* trainer -> new _tensor_to function to support final data conversions i.e. cpu to gpu mem if needed for tuple types for both outputs
	     and inputs

with all of this the error for multi-out was very bad, the hope is that multi-in is better?

+++ metrics/loss.py
@@ -22,4 +22,4 @@ def mse_and_nll_loss(outputs: tuple, targets: tuple):
     target_hpe, target_cidx = targets
 
     # combine the two losses with eual weights, TODO: adjust loss weights in future...
-    return F.mse_loss(output_hpe, target_hpe) + F.nll_loss(output_probvect, target_cidx)
+    return F.mse_loss(output_hpe, target_hpe) + F.nll_loss(output_probvect, target_cidx.long())
